rename the test/Main.hs to test/Spec.hs, only the Module in Spec.hs is Main will be fine for testing!  Change consine type from Double to Folat so that the test will be success, otherwise it will be stucked on 1 <> 1.00000000000002 thing like that, too much accurancy

+++ src/Eris/Compute/Similarity.hs
@@ -1,13 +1,13 @@
 module Eris.Compute.Similarity
     (
-      consine 
+      consine
     ) where
 import qualified Data.HashMap.Strict as Map
 import Data.Maybe
 
 import Eris.Meta.DataStructure
 
-consine :: [Double] -> [Double] -> Double
+consine :: [Float] -> [Float] -> Float
 consine v1 v2 = dot v1 v2 / (elen v1 * elen v2)
     where
         dot a b = sum $ zipWith (*) a b
